+++ test/fixtures/cookbooks/test/recipes/source.rb
@@ -8,6 +8,7 @@ end
 haproxy_install 'source' do
   source_url node['haproxy']['source_url']
   source_checksum node['haproxy']['source_checksum']
+  source_version node['haproxy']['source_version']
   use_pcre '1'
   use_openssl '1'
   use_zlib '1'